changes inncontrollers

+++ gestionprojet/src/main/java/com/este/gestionprojet/repository/UtilisateurRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UtilisateurRepo extends JpaRepository<Utilisateur, Long> {
+public interface UtilisateurRepo extends JpaRepository<Utilisateur, Integer> {
 }
